pyproject.toml
[project]
name = "searchai-experimental"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.1",
    "ruff>=0.11.9",
    "requests>=2.32.3",
    "python-dotenv>=1.1.0",
    "crewai-tools>=0.45.0",
    "openai>=1.75.0",
    "groq>=0.24.0",
    "django-stubs>=5.2.0",
    "extended-mypy-django-plugin>=0.8.0",
    "django-htmx>=1.23.0",
    "beautifulsoup4>=4.13.4",
    "lxml[html-clean]>=4.9.2",
    "newspaper3k>=0.2.8",
    "trafilatura>=2.0.0",
    "celery>=5.5.2",
    "tiktoken>=0.7.0",
    "mypy>=1.15.0",
]

[project.scripts]
searchai = "agent.cli:app"
searchai-quiet = "agent.cli_quiet:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent", "app", "project"]

[tool.hatch.envs.default]
dev-mode = true
features = []

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "C4", "COM", "Q", "SIM", "UP", "W"]
ignore = [
    "PLW2901", # Redefined outer name
    "B904",    # Within an `except` clause, raise without `from`
    "E501",    # Line too long
]

# Warning suppression for Pydantic
[tool.ruff.lint.per-file-ignores]
"pydantic/_internal/_generate_schema.py" = ["PLW0711"]

[tool.ruff.lint.isort]
known-first-party = ["agent", "app", "project"]

[tool.mypy]
exclude = "\\.venv"
plugins = ["extended_mypy_django_plugin.main"]
python_version = "3.12"
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unreachable = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
disallow_untyped_decorators = false
no_implicit_optional = true
strict_optional = true

[tool.django-stubs]
django_settings_module = "project.settings"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "manage"
ignore_errors = true

[[tool.mypy.overrides]]
module = "crewai.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "crewai_tools.*"
ignore_errors = true
